When the spinner is hidden, hide the stub view.
The problem with returning `<View />` is that it still counts as a
flexbox child and thus takes up space in the layout. Adding and
removing the spinner in that situation causes the layout to jump
around.

Returning null would also be a reasonable solution, but in practice
it seemed to cause a layout bug on Android wherein the spinner overlay
wouldn't fully disappear.

Returning a hidden view solves both of these problems.